+++ test/integration_test.js
@@ -29,13 +29,6 @@ AnnotatingPlugin.prototype.build = function() {
   }
 }
 
-NoopPlugin.prototype = Object.create(Plugin.prototype)
-NoopPlugin.prototype.constructor = NoopPlugin
-function NoopPlugin() {
-  Plugin.apply(this, arguments)
-}
-NoopPlugin.prototype.build = function() {}
-
 var build = function(builder) {
   return RSVP.Promise.resolve()
     .then(function() {
@@ -68,7 +61,7 @@ describe('integration test', function(){
     })
 
     it('sets description', function() {
-      var node = new NoopPlugin([], {
+      var node = new AnnotatingPlugin([], {
         name: 'SomePlugin',
         annotation: 'some annotation'
       })
@@ -134,53 +127,6 @@ describe('integration test', function(){
       it('is persistent when persistentOutput is true', function() {
         return expect(isPersistent({ persistentOutput: true })).to.eventually.equal(true)
       })
-    })
-  })
-})
-
-
-describe('unit tests', function() {
-  it('toString', function() {
-    expect(new NoopPlugin([]) + '').to.equal('[NoopPlugin]')
-    expect(new NoopPlugin([], { name: 'FooPlugin' }) + '').to.equal('[FooPlugin]')
-    expect(new NoopPlugin([], { annotation: 'some note' }) + '').to.equal('[NoopPlugin: some note]')
-  })
-
-  describe('usage errors', function() {
-    it('requires the base constructor to be called (super)', function() {
-      TestPlugin.prototype = Object.create(Plugin.prototype)
-      TestPlugin.prototype.constructor = TestPlugin
-      function TestPlugin() { /* no Plugin.apply(this, arguments) here */ }
-      TestPlugin.prototype.build = function() {}
-
-      return expect(build(new Builder_0_16_3(new TestPlugin)))
-        .to.be.rejectedWith(Error, /must call the superclass constructor/)
-    })
-
-    it('disallows overriding read, cleanup, and rebuild', function() {
-      var prohibitedNames = ['read', 'rebuild', 'cleanup']
-      for (var i = 0; i < prohibitedNames.length; i++) {
-        var BadPlugin = function BadPlugin() {
-          Plugin.apply(this, arguments)
-        }
-        BadPlugin.prototype = Object.create(Plugin.prototype)
-        BadPlugin.prototype.constructor = BadPlugin
-        BadPlugin.prototype.build = function() {}
-        BadPlugin.prototype[prohibitedNames[i]] = function() {}
-
-        expect(function() { new BadPlugin([]) })
-          .to.throw(/For compatibility, plugins must not define/)
-      }
-    })
-
-    it('checks that the inputNodes argument is an array', function() {
-      expect(function() { new AnnotatingPlugin('notAnArray') })
-        .to.throw(/Expected an array/)
-    })
-
-    it('provides a helpful error message on missing `new`', function() {
-      expect(function() { AnnotatingPlugin([]) })
-        .to.throw(/Missing `new`/)
     })
   })
 })