+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # master
 
+* Add `sourceDirectories` feature flag, which introduces `pluginInterface.nodeType`
+* Allow for calling `__broccoliGetInfo__()` without argument
+
 # 1.1.0
 
 * Add `persistentOutput` flag